renamed scramble component files
