+++ src/components/scramble/index.ts
@@ -1,3 +1,3 @@
-import Scramble from "./component";
+import Scramble from "./scramble";
 
 export default Scramble;
